// types.ts
export enum ExerciseType {
  FILL_IN_THE_BLANK = 'Completar los espacios',
  MULTIPLE_CHOICE = 'Opción múltiple',
  SENTENCE_ORDERING = 'Ordenar las frases',
}

export enum Difficulty {
  A1 = 'Principiante (A1)',
  A2 = 'Principiante alto (A2)',
  B1 = 'Intermedio (B1)',
  B2 = 'Intermedio alto (B2)',
  C1 = 'Avanzado (C1)',
}